Enable spell check toggle
When I'm writing a commit message I want it. As I use Vim usually for
that purpose I have it enabled by default; but sometimes I edit code
with Vim, and it's really ugly then. So I want and easy way to switch it
off, and that's what I add in this commit.